restructure mathLib, mathLib_vector
