+++ magpylib/_lib/fields/Moment_Dipole.py
@@ -22,7 +22,7 @@
 # page at https://www.github.com/magpylib/magpylib/issues.
 # -------------------------------------------------------------------------------
 
-from magpylib._lib.mathLibPrivate import fastSum3D
+from magpylib._lib.mathLib import fastSum3D
 from numpy import pi, array, NaN
 from warnings import warn
