+++ tests/test_privateMath.py
@@ -1,6 +1,6 @@
-from magpylib._lib.mathLibPrivate import getPhi,fastSum3D, fastNorm3D, arccosSTABLE, fastCross3D
-from magpylib._lib.mathLibPrivate import Qmult, Qnorm2, Qunit, Qconj, getRotQuat, angleAxisRotation
-from magpylib._lib.mathLibPrivate import elliptic, ellipticK, ellipticE, ellipticPi
+from magpylib._lib.mathLib import getPhi,fastSum3D, fastNorm3D, arccosSTABLE, fastCross3D
+from magpylib._lib.mathLib import Qmult, Qnorm2, Qunit, Qconj, getRotQuat, angleAxisRotation
+from magpylib._lib.mathLib import elliptic, ellipticK, ellipticE, ellipticPi
 from numpy import pi,array
